moved user collection related mongo functions to user_mongo_client
